% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosmxSXE.R
\name{readCosmxSXE}
\alias{readCosmxSXE}
\title{Load data from a Nanostring CosMx experiment}
\usage{
readCosmxSXE(
  dirName = dirName,
  returnType = "SPE",
  countMatPattern = "exprMat_file.csv",
  metaDataPattern = "metadata_file.csv",
  coordNames = c("CenterX_global_px", "CenterY_global_px"),
  loadFovPos = FALSE,
  fovPosPattern = "fov_positions_file.csv",
  altExps = NULL,
  addParquetPaths = FALSE,
  ...
)
}
\arguments{
\item{dirName}{a directory path to CosMx download that contains files of interest.}

\item{returnType}{option of \code{"SPE"} or \code{"SCE"}, stands for
\code{SpatialExperiment} or \code{SingleCellExperiment} object. Default value \code{"SPE"}}

\item{countMatPattern}{a filename pattern for the count matrix. Default value is
\code{"exprMat_file.csv"}, and there is no need to change.}

\item{metaDataPattern}{a filename pattern for the metadata .csv file that
contains spatial coords. Default value is \code{"metadata_file.csv"}, and
there is no need to change.}

\item{coordNames}{a vector of two strings specify the spatial coord names.
Default value is \code{c("CenterX_global_px", "CenterY_global_px")}, and
there is no need to change.}

\item{loadFovPos}{to merge fov_position_list.csv to \code{colData(sxe)} or not.
Default is FALSE.}

\item{fovPosPattern}{.csv pattern of fov_position_list.csv files in the raw download.
Default value is \code{"fov_positions_file.csv"}.}

\item{altExps}{gene names contains these strings will be moved to \code{altExps(sxe)}
as list of separate sxe. Default is NULL. Suggest \code{c(negprobe="^Neg", falsecode="^Sys")}.}

\item{addParquetPaths}{to add parquet paths to \code{metadata(sxe)} or not. If TRUE,
transcripts and polygon .csv files will be converted to .parquet, and the paths
will be added. If, for instance, no polygon file is available, and only transcript
file is available, please set this argument to TRUE and adjust loadPolygon = FALSE
in the \code{...} argument.}

\item{...}{extra parameters to pass to addParquetPathsCosMx(), including
\code{loadTx}, \code{txMetaNames}, \code{txPattern}, \code{loadPolygon}, \code{polygonMetaNames},
\code{polygonPattern}.}
}
\value{
a \code{\link{SpatialExperiment}} or a \code{\link{SingleCellExperiment}} object
}
\description{
Creates a \code{\link{SpatialExperiment}} from the downloaded unzipped CosMx
directory for Nanostring CosMx spatial gene expression data.
}
\details{
The constructor assumes the downloaded unzipped CosMx folder has the following
structure, with two mandatory files:
CosMx_unzipped/optional_default_folder/ \cr
· | — *_exprMat_file.csv \cr
· | — *_metadata_file.csv \cr

Optional files to add to the metadata() as a list of paths (will be converted to parquet):
· | — *_tx_file.csv \cr
· | — *_polygons.csv \cr
· | — *_fov_positions_file.csv \cr
See addParquetPathsCosmx()
}
\examples{
# A relatively small data download can be from:
# https://nanostring.com/resources/smi-ffpe-dataset-lung9-rep1-data/


# A mock counts and mock metadata with spatial location generated for a 8 genes by 
# 9 cells object is in /extdata: 

cospath <- system.file(
  file.path("extdata", "CosMx_small"),
  package = "SpatialExperimentIO")
  
list.files(cospath)

# One of the following depending on your output (`SPE` or `SCE`) requirement.
cos_spe <- readCosmxSXE(dirName = cospath)
cos_sce <- readCosmxSXE(dirName = cospath, returnType = "SCE")
cos_spe <- readCosmxSXE(dirName = cospath, addParquetPaths = TRUE)
cos_spe <- readCosmxSXE(dirName = cospath, addParquetPaths = TRUE, loadPolygon = FALSE)


}
\author{
Yixing Estella Dong
}
