% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosmxSPE.R
\name{readCosmxSPE}
\alias{readCosmxSPE}
\title{Load data from a Nanostring CosMx experiment}
\usage{
readCosmxSPE(
  dirname = dirname,
  countmatfpattern = "exprMat_file.csv",
  metadatafpattern = "metadata_file.csv",
  coord_names = c("CenterX_global_px", "CenterY_global_px")
)
}
\arguments{
\item{dirname}{a directory path to CosMx download that contains files of interest.}

\item{countmatfpattern}{a filename pattern for the count matrix. Default value is
\code{"exprMat_file.csv"}, and there is no need to change.}

\item{metadatafpattern}{a filename pattern for the metadata .csv file that
contains spatial coords. Default value is \code{"metadata_file.csv"}, and
there is no need to change.}

\item{coord_names}{a vector of two strings specify the spatial coord names.
Default value is \code{c("CenterX_global_px", "CenterY_global_px")}, and
there is no need to change.}
}
\value{
a \code{\link{SpatialExperiment}} object
}
\description{
Creates a \code{\link{SpatialExperiment}} from the downloaded unzipped CosMx
directory for Nanostring CosMx spatial gene expression data.
}
\details{
The constructor assumes the downloaded unzipped CosMx folder has the following
structure, with two mandatory files:
CosMx_unzipped/optional_default_folder/ \cr
· | — *_exprMat_file.csv \cr
· | — *_metadata_file.csv \cr
}
\examples{
\dontrun{
# Data download is from: 
# https://nanostring.com/resources/smi-ffpe-dataset-lung9-rep1-data/

# Here as an example, we have downsized the count matrix and meta data file 
# to only 453 randomly selected cells and 20 genes, and necessary columns 
# of the metadata. 

cospath <- system.file(
  file.path("extdata", "NanostringCosMx"),
  package = "SpatialExperiment")
  
list.files(cospath)

cos_spe <- readCosmxSPE(dirname = cospath, 
                        countmatfpattern = "exprMat_file.csv", 
                        metadatafpattern = "metadata_file.csv", 
                        coord_names = c("CenterX_global_px",
                                        "CenterY_global_px"))
}

}
\author{
Estella Yixing Dong
}
