% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStarmapplusSXE.R
\name{readStarmapplusSXE}
\alias{readStarmapplusSXE}
\title{Load data from a STARmap PLUS experiment}
\usage{
readStarmapplusSXE(
  dirname = dirname,
  return_type = "SPE",
  countmatfpattern = "raw_expression_pd.csv",
  metadatafpattern = "spatial.csv",
  coord_names = c("X", "Y", "Z")
)
}
\arguments{
\item{dirname}{a directory path to CosMx download that contains files of interest.}

\item{return_type}{option of \code{"SPE"} or \code{"SCE"}, stands for
\code{SpatialExperiment} or \code{SingleCellExperiment} object. Default value \code{"SPE"}}

\item{countmatfpattern}{a filename pattern for the count matrix. Default value is
\code{"exprMat_file.csv"}, and there is no need to change.}

\item{metadatafpattern}{a filename pattern for the metadata .csv file that
contains spatial coords. Default value is \code{"metadata_file.csv"}, and
there is no need to change.}

\item{coord_names}{a vector of two strings specify the spatial coord names.
Default value is \code{c("CenterX_global_px", "CenterY_global_px")}, and
there is no need to change.}
}
\value{
a \code{\link{SpatialExperiment}} or a \code{\link{SingleCellExperiment}} object
}
\description{
Creates a \code{\link{SpatialExperiment}} from the downloaded unzipped CosMx
directory for Nanostring CosMx spatial gene expression data.
}
\details{
The constructor assumes the downloaded unzipped CosMx folder has the following
structure, with two mandatory files:
CosMx_unzipped/optional_default_folder/ \cr
· | — *_exprMat_file.csv \cr
· | — *_metadata_file.csv \cr
}
\examples{
\dontrun{
# Data download is from: 
# https://zenodo.org/records/8327576

TODO: not this ~ ---------------------------------
starpath <- system.file(
  file.path("extdata", "STARmap_PLUS"),
  package = "SpatialExperimentIO")
--------------------------------------------------
  
TODO: But something like this! -------------------
starpath <- "path/to/a/defined/folder/"

eh <- ExperimentHub()
query(eh, c("scSpatialExperimentData", "STARmap_PLUS"))

count.dat <- eh[["EH7543"]]
write.csv(count.dat, file.path(starpath, "raw_expression_pd.csv")
col.dat <- eh[["EH7544"]]
write.csv(col.dat, file.path(starpath, "spatial.csv")
--------------------------------------------------

list.files(starpath)

# One of the following depending on your output (`SPE` or `SCE`) requirement.
star_spe <- readStarmapplusSXE(dirname = starpath)
star_sce <- readStarmapplusSXE(dirname = starpath, return_type = "SCE")

}

}
\author{
Estella Yixing Dong
}
