% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXeniumSXE.R
\name{readXeniumSXE}
\alias{readXeniumSXE}
\title{Load data from a 10x Geonomics Xenium experiment}
\usage{
readXeniumSXE(
  dirname,
  return_type = "SPE",
  countfname = "cell_feature_matrix.h5",
  coordfpattern = "cells.csv.gz",
  coord_names = c("x_centroid", "y_centroid")
)
}
\arguments{
\item{dirname}{a directory path to Xenium Output Bundle download that contains
files of interest.}

\item{return_type}{option of \code{"SPE"} or \code{"SCE"}, stands for
\code{SpatialExperiment} or \code{SingleCellExperiment} object. Default value \code{"SPE"}}

\item{countfname}{a folder directory or the h5 file for the count matrix.
Default value is \code{"cell_feature_matrix.h5"}, alternative value is
\code{"cell_feature_matrix"} that takes a bit longer. The count matrix is
read in and stored in a \code{SingleCellExperiment} object, using
\code{DropletUtils::read10xCounts()}}

\item{coordfpattern}{a filename pattern of the zipped .csv file that
contains spatial coords. Default value is \code{"cells.csv.gz"}, and there is no
need to change.}

\item{coord_names}{a vector of two strings specify the spatial coord names.
Default value is \code{c("x_centroid", "y_centroid")}, and there is no need to change.}
}
\value{
a \code{\link{SpatialExperiment}} or a \code{\link{SingleCellExperiment}} object
}
\description{
Creates a \code{\link{SpatialExperiment}} from the downloaded unzipped Xenium
Output Bundle directory for 10x Genomics Xenium spatial gene expression data.
}
\details{
The constructor assumes the downloaded unzipped Xenium Output Bundle has the
following structure, with mandatory file of cells.csv.gz and either folder
/cell_feature_matrix or .h5 file cell_feature_matrix.h5:
Xenium_unzipped \cr
· | — cell_feature_matrix.h5 \cr
· | — cell_feature_matrix \cr
· · | - barcodes.tsv.gz \cr
· · | - features.tsv.gz \cr
· · | - matrix.mtx.gz \cr
· | — cells.csv.gz \cr
}
\examples{
\dontrun{
# Data download is from: 
# https://cf.10xgenomics.com/samples/xenium/1.0.2/Xenium_V1_FF_Mouse_Brain_
# Coronal_Subset_CTX_HP/Xenium_V1_FF_Mouse_Brain_Coronal_Subset_CTX_HP_outs.zip

xepath <- system.file(
  file.path("extdata", "10xXenium"),
  package = "SpatialExperimentIO")
  
list.files(xepath)

# One of the following depending on your input (.h5 or folder) and output 
# (`SPE` or `SCE`) requirement.
xe_spe <- readXeniumSXE(dirname = xepath)
xe_spe <- readXeniumSXE(dirname = xepath, countfname = "cell_feature_matrix")
xe_sce <- readXeniumSXE(dirname = xepath, return_type = "SCE")

# Subset to no control genes, and the same needed for `xe_sce` if read in as 
# `SCE`.
xe_spe <- xe_spe[rowData(xe_spe)$Type == "Gene Expression"]

}
}
\author{
Estella Yixing Dong
}
